FIX: selfRedirectCanonicalizer in CompositeAccessPoint is ignored and causes infinite redirect loop with self-redirect captures. 	urlToKey() method accesses selfRedirectCanonicalizer directly. It should have accessed it thorough the getter method. 	Move exactSchemeMatch and selfRedirectCanonicalizer back to AccessPont to prevent this sort of errors in the future. AccessPointBase now accesses those fields through abstract getter methods.
ARI-5234
